# api-gateway/internal/http/docs/swagger.yaml
definitions:
...
  genproto.DraftCreateUpdateBody:
    properties:
      attachment_ids:
        items:
          type: string
        type: array
      body:
        type: string
      receivers:
        $ref: '#/definitions/genproto.Receivers'
      subject:
        type: string
    type: object
  genproto.DraftCreateUpdateReq:
    properties:
      body:
        $ref: '#/definitions/genproto.DraftCreateUpdateBody'
      sender_id:
        type: string
    type: object
...
  genproto.MessageSendBCC:
    properties:
      email:
        items:
          type: string
        type: array
    type: object
  genproto.MessageSendCC:
    properties:
      email:
        items:
          type: string
        type: array
    type: object
  genproto.MessageSendTo:
    properties:
      email:
        items:
          type: string
        type: array
    type: object
  genproto.MessageSentRes:
    properties:
      failed_emails:
        items:
          type: string
        type: array
      total_failed:
        type: integer
      total_sent:
        type: integer
    type: object
...
  genproto.Receivers:
    properties:
      bcc:
        $ref: '#/definitions/genproto.MessageSendBCC'
      cc:
        $ref: '#/definitions/genproto.MessageSendCC'
      to:
        $ref: '#/definitions/genproto.MessageSendTo'
    type: object
...
  genproto.Void:
    type: object
...
info:
  contact: {}
paths:
...
  /draft:
    post:
      consumes:
      - application/json
      description: Creates a new draft message associated with the authenticated user.
      parameters:
      - description: Draft creation request
        in: body
        name: draft
        required: true
        schema:
          $ref: '#/definitions/genproto.DraftCreateUpdateBody'
      produces:
      - application/json
      responses:
        "201":
          description: Draft created successfully
          schema:
            $ref: '#/definitions/genproto.Void'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new draft
      tags:
      - 05-Draft
  /draft/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a draft message with the provided ID.
      parameters:
      - description: Draft ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Draft deleted successfully
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Draft not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a draft
      tags:
      - 05-Draft
    put:
      consumes:
      - application/json
      description: Updates a draft message with the provided ID.
      parameters:
      - description: Draft ID
        in: path
        name: id
        required: true
        type: string
      - description: Draft update request
        in: body
        name: draft
        required: true
        schema:
          $ref: '#/definitions/genproto.DraftCreateUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Draft updated successfully
          schema:
            $ref: '#/definitions/genproto.Void'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Draft not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update an existing draft
      tags:
      - 05-Draft
  /draft/{id}/send:
    post:
      consumes:
      - application/json
      description: Sends a draft message with the provided ID.
      parameters:
      - description: Draft ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Draft sent successfully
          schema:
            $ref: '#/definitions/genproto.MessageSentRes'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Draft not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Send a draft
      tags:
      - 05-Draft
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...